重构代码，替换 Handler 为 utils，整理引用
