+++ src/plugins/Main.py
@@ -17,7 +17,7 @@ from nonebot.adapters.onebot.v11 import (
     Bot,
     )
 from nonebot.matcher import Matcher
-import Handler
+import utils
 from nonebot.rule import to_me,is_type,Rule
 from nonebot.params import CommandArg
 from nonebot.permission import SUPERUSER
@@ -52,7 +52,7 @@ async def chek_boss(event:MessageEvent):
 
 
 async def chek_add_welcome(event:GroupIncreaseNoticeEvent):
-    Welcome_Dict = Handler.handle_json(Welcome_Path, 'r')
+    Welcome_Dict = utils.handle_json(Welcome_Path, 'r')
     Group_id = str(event.group_id)
     logger.info(f"检查群 {Group_id} 的欢迎配置，当前数据：{Welcome_Dict.get(Group_id)}")
     if Welcome_Dict.get(Group_id,False):
@@ -64,7 +64,7 @@ async def chek_add_welcome(event:GroupIncreaseNoticeEvent):
 
 async def ChekGroupMemberChange(event:GroupDecreaseNoticeEvent):
     try:
-        Data = Handler.handle_json(f'{path}/GroupMemberChange.json', "r")
+        Data = utils.handle_json(f'{path}/GroupMemberChange.json', "r")
         logger.info(f"检查群 {event.group_id} 的退群通知开关，当前状态：{Data.get(str(event.group_id))}")
         return Data.get(str(event.group_id), False)
     except Exception as e:
@@ -72,7 +72,7 @@ async def ChekGroupMemberChange(event:GroupDecreaseNoticeEvent):
         return False
 
 async def add_group_switch(event:GroupRequestEvent):
-    Data = Handler.handle_json(f'{path}/add_group_switch.json', "r")
+    Data = utils.handle_json(f'{path}/add_group_switch.json', "r")
     logger.info(Data.get(str(event.group_id), False))
     logger.info(event.sub_type == "add")
     return Data.get(str(event.group_id), False) and event.sub_type == "add"
@@ -107,14 +107,14 @@ AT_Time = time.time()
 
 
 @Test.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Test_Function(matcher: Matcher,bot: Bot,event:MessageEvent, args: Message = CommandArg()):
     logger.info(await bot.get_group_member_info(group_id=event.group_id,user_id=event.self_id))
     raise RuntimeError("Test Function Error")  # 测试用的异常抛出
 
 
 @Poke_Check.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def PC_Function(matcher:Matcher,event: PokeNotifyEvent):
     global Poke_Count, Time_Count, Send
     if Poke_Count >= 3 and time.time()-Time_Count <= 120:
@@ -129,7 +129,7 @@ async def PC_Function(matcher:Matcher,event: PokeNotifyEvent):
                 pass
     else:
         Send = True
-        Text_List = Handler.handle_json(Poke_Path, 'r')
+        Text_List = utils.handle_json(Poke_Path, 'r')
             
         Poke_Count += 1
         Time_Count = time.time()
@@ -137,7 +137,7 @@ async def PC_Function(matcher:Matcher,event: PokeNotifyEvent):
         await matcher.finish(f"{Choise_Text}")
 
 @Tarot.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Tarot_Function(matcher:Matcher,event: MessageEvent):
     get = requests.get("https://oiapi.net/API/Tarot").json()
     if get['code'] != 1:
@@ -156,10 +156,10 @@ async def Tarot_Function(matcher:Matcher,event: MessageEvent):
 
 Add_Welcome = on_notice(rule=is_type(GroupIncreaseNoticeEvent) & Rule(chek_add_welcome),priority=1,block=True)
 @Add_Welcome.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Welcome(matcher:Matcher,event=GroupIncreaseNoticeEvent):
     if event.user_id == event.self_id:await matcher.finish()
-    Welcome_Dict = Handler.handle_json(Welcome_Path, 'r')
+    Welcome_Dict = utils.handle_json(Welcome_Path, 'r')
     Group_id = str(event.group_id)
     if Welcome_Dict.get(Group_id,False):
         Group_Dict = Welcome_Dict[f'{Group_id}']
@@ -175,16 +175,16 @@ async def Welcome(matcher:Matcher,event=GroupIncreaseNoticeEvent):
 
 SelfJoinGroupWelcome = on_notice(rule=is_type(GroupIncreaseNoticeEvent),priority=1,block=True)
 @SelfJoinGroupWelcome.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def SelfJoinGroupWelcome_Function(matcher:Matcher,event=GroupIncreaseNoticeEvent):
     if event.user_id == event.self_id:await Matcher.finish("嗷呜~感谢您使用凌辉Bot~我可以给你提供Furry的相关功能，以及一些其他的功能哦~\n如果您想要了解更多功能，请输入“菜单”来获取帮助信息哦~\n如果您担心与群里其他Bot的命令冲突，可以通过“凌辉菜单”来使用菜单哦www~希望您能喜欢我~\n如果您有任何问题，请随时联系管理员[1097740481]哦~")
     else:
         pass
 
 @Change_Welcome.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Change_Welcome_Function(matcher:Matcher,event:GroupMessageEvent,Recivce:MessageEvent):
-    Dict = Handler.handle_json(Welcome_Path, 'r')
+    Dict = utils.handle_json(Welcome_Path, 'r')
     Group_id = str(event.group_id)
     args = Recivce.get_message()
     if "开" in str(args):
@@ -211,14 +211,14 @@ async def Change_Welcome_Function(matcher:Matcher,event:GroupMessageEvent,Recivc
         elif Mode==None:
             Mode_Text = "未启动"
         await matcher.finish(MessageSegment.reply(event.message_id)+f"当前群聊的入群欢迎状态为：{Mode_Text}\n新人入群欢迎文本：{Text}")
-    Handler.handle_json(Welcome_Path, 'w', Dict)
+    utils.handle_json(Welcome_Path, 'w', Dict)
     await matcher.finish(MessageSegment.reply(event.message_id)+"操作成功完成。")
 
 @Change_Welcome_Text.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def CWT_Function(matcher:Matcher,event:MessageEvent,args:Message = CommandArg()):
     args = str(args)
-    Dict = Handler.handle_json(Welcome_Path, 'r')
+    Dict = utils.handle_json(Welcome_Path, 'r')
     Group = str(event.group_id)
     Text = ""
     if not Dict.get(Group,False):
@@ -234,25 +234,25 @@ async def CWT_Function(matcher:Matcher,event:MessageEvent,args:Message = Command
     if args == "":
         Dict[Group]['Text'] = "新人记得给群主早上请安晚上侍寝（bushi\n欢迎新成员加入本群！凌辉Bot欢迎您~\nWelcome new members to join this family! Linghui Bot welcomes you~"
         Text_1 = "由于未找到对应的文本，所以本次操作将会使用默认文本来进行入群欢迎~"
-    Handler.handle_json(Welcome_Path, 'w', Dict)
+    utils.handle_json(Welcome_Path, 'w', Dict)
     await matcher.finish(MessageSegment.reply(event.message_id)+f"{Text}{Text_1}")        
 
 @AWord.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def AWord_Function(matcher:Matcher,event=MessageEvent,args:Message = CommandArg()):
     if args.extract_plain_text():await matcher.finish()    # 若消息后面存在文本则不响应
-    Dict = Handler.handle_json(AWord_Path, 'r')
+    Dict = utils.handle_json(AWord_Path, 'r')
         
     Result = Dict[rd.randint(0,len(Dict)-1)]
     await matcher.finish(MessageSegment.reply(event.message_id)+f"“{Result}”")
 
 #签到触发器与实现
 @Sign_in.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Sign_in_Function(matcher:Matcher,event:MessageEvent,GroupEvent:GroupMessageEvent,args:Message=CommandArg()):
     if args.extract_plain_text():await matcher.finish()    # 若消息后面存在文本则不响应
     # 打开文件并写入Sign_Dict字典
-    Sign_Dict = Handler.handle_json(Sign_in_Path, 'r')
+    Sign_Dict = utils.handle_json(Sign_in_Path, 'r')
 
     # 获取触发人QQ号和群聊号
     User,Group = event.user_id,GroupEvent.group_id
@@ -305,8 +305,8 @@ async def Sign_in_Function(matcher:Matcher,event:MessageEvent,GroupEvent:GroupMe
     Group_User_List[f"{User}"] = New_Sign_User_Dict
     Sign_Dict[f'{Group}'] = New_Sign_Group_Dict
     # 将构建完成的信息写入本地json文件进行保存
-    Handler.handle_json(Sign_in_Path, 'w', Sign_Dict)
-    AwordDict = Handler.handle_json(AWord_Path, 'r')
+    utils.handle_json(Sign_in_Path, 'w', Sign_Dict)
+    AwordDict = utils.handle_json(AWord_Path, 'r')
     Result = AwordDict[rd.randint(0,len(AwordDict)-1)]
     # 判断：调用是否出现“好久不见”字样
     if "好久不见" in str(event.message):
@@ -324,7 +324,7 @@ async def Sign_in_Function(matcher:Matcher,event:MessageEvent,GroupEvent:GroupMe
     await matcher.finish(MessageSegment.reply(event.message_id)+f"{Text}签到成功。您本月在本群中已签到{User_Count}次，今天在本群中排名第{Group_Count}位。\n——————\n“{Result}”")
 
 @btfrk.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def wc_btfrk(bot:Bot,matcher:Matcher,event:MessageEvent):
     match = re.search(r"我是(.+)控", str(event.original_message))
     members = await bot.get_group_member_list(group_id=event.group_id)
@@ -348,10 +348,10 @@ async def wc_btfrk(bot:Bot,matcher:Matcher,event:MessageEvent):
     await bot.send(event, message)
 
 @Exit_Change.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def Change_Exit_Function(matcher:Matcher,event:MessageEvent,args:Message=CommandArg()):
     args = str(args)
-    Data = Handler.handle_json(f'{path}/GroupMemberChange.json', "r")
+    Data = utils.handle_json(f'{path}/GroupMemberChange.json', "r")
     Temp_1,write = "",None
     if "开" in args:
         write = True
@@ -368,7 +368,7 @@ async def Change_Exit_Function(matcher:Matcher,event:MessageEvent,args:Message=C
         await matcher.finish(MessageSegment.reply(event.message_id)+f"{Text}~\n输入“退群提示关”或“退群提示开”来更改功能开关")
     logger.info(args)
     Data[f'{event.group_id}']=write
-    Handler.handle_json(f'{path}/GroupMemberChange.json', "w", Data)
+    utils.handle_json(f'{path}/GroupMemberChange.json', "w", Data)
     await matcher.finish(MessageSegment.reply(event.message_id)+f"好~凌辉Bot已经{Temp}了本群的退群提示w~{Temp_1}")
 
 # 检查群成员减少事件
@@ -379,7 +379,7 @@ GroupExitMember = on_notice(
 )
 
 @GroupExitMember.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def handle_group_decrease(event: GroupDecreaseNoticeEvent, bot: Bot, matcher: Matcher):
     logger.info(f"触发退群事件：{event.dict()}")
     # 处理 Bot 自己被踢的情况
@@ -450,12 +450,12 @@ async def chekFriendLike(event:NoticeEvent):
 Like_Friend = on_notice(rule=is_type(NoticeEvent)&chekFriendLike)
 
 @Like_Friend.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def LF_Function(matcher:Matcher,event:NoticeEvent):
     from nonebot import get_bot
     bot = get_bot()
     Data = event.model_dump()
-    Dict = Handler.handle_json(f"{path}/Friend_Like.json", 'r')
+    Dict = utils.handle_json(f"{path}/Friend_Like.json", 'r')
     user_id = Data['operator_id']
     now = dt.now()
     day = now.day
@@ -478,7 +478,7 @@ async def LF_Function(matcher:Matcher,event:NoticeEvent):
     Text = f"owo？是你给我点赞了嘛~谢谢~\n凌辉也给你点赞哦~（如果已经点过了那就不点了呐~嘻嘻ww）"
     Dict[f"{user_id}"]['Model'] = False
     Dict[f"{user_id}"]['Time'] = day
-    Handler.handle_json(f"{path}/Friend_Like.json", 'w', Dict)
+    utils.handle_json(f"{path}/Friend_Like.json", 'w', Dict)
     try:
         if user_id != 1097740481:
             await bot.send_private_msg(user_id=user_id,message=Text)
@@ -491,7 +491,7 @@ async def LF_Function(matcher:Matcher,event:NoticeEvent):
 Add_Friend = on_request(rule=is_type(FriendRequestEvent),priority=1, block=True)
 
 @Add_Friend.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def AF_Function(bot:Bot,matcher:Matcher,event:FriendRequestEvent):
     request_type,user_id,flag,comment = event.request_type,event.user_id,event.flag,event.comment
     stranger_info = await bot.get_stranger_info(user_id=int(user_id))
@@ -557,17 +557,17 @@ async def Eat_Function(matcher:Matcher,event:MessageEvent,bot:Bot,args:Message =
     await matcher.finish(MessageSegment.reply(event.message_id)+f"{a['mealwhat']}\n要不{Select}吧！")
 
 @add_group.handle()
-@Handler.handle_errors
+@utils.handle_errors
 async def handle_add_group(matcher:Matcher,bot:Bot,event:GroupRequestEvent):
     User = await bot.get_stranger_info(group_id=event.group_id,user_id=event.user_id)
     User = User["nick"]
-    Data = Handler.handle_json(f'{path}/add_group_switch.json', 'r')
+    Data = utils.handle_json(f'{path}/add_group_switch.json', 'r')
     if Data.get(str(event.group_id),False):
         flag_list = Data[str(event.group_id)]
     else:
         flag_list = [event.flag]
     # Dict = {event.group_id:flag_list}
-    # Handler.load_json(f'{path}/add_group_switch.json','w',Dict)
+    # utils.load_json(f'{path}/add_group_switch.json','w',Dict)
     comment = event.comment
     if comment == "":
         comment = "未填写入群理由"
@@ -578,9 +578,9 @@ async def handle_add_group(matcher:Matcher,bot:Bot,event:GroupRequestEvent):
     await matcher.finish(f"似乎有人想要加入我们awa...\n请求类型：{event.request_type}\n子类型：{event.sub_type}\n申请人信息：{User}[{event.user_id}]\n进群理由:{event.comment}\n要同意此人入群嘛awa？\n可以通过“允许加群{event.flag}”或“拒绝加群{event.flag}”来处理此请求（请在Bot为群管理员时进行操作~")
 
 @switch_add_group.handle()
-async def handler_switch_add_group(matcher:Matcher,event:MessageEvent,args:Message=CommandArg()):
+async def utils_switch_add_group(matcher:Matcher,event:MessageEvent,args:Message=CommandArg()):
     args = str(args)
-    Data = Handler.handle_json(f'{path}/add_group_switch.json', 'r')
+    Data = utils.handle_json(f'{path}/add_group_switch.json', 'r')
     if "开" in args:
         write = True
         Temp = "打开"
@@ -594,7 +594,7 @@ async def handler_switch_add_group(matcher:Matcher,event:MessageEvent,args:Messa
                 Text = "当前Bot的入群检测状态为：开启"
         await matcher.finish(MessageSegment.reply(event.message_id)+f"{Text}~\n输入“入群检测开”或“入群检测关”来更改功能开关")
     Data[str(event.group_id)] = write
-    Handler.handle_json(f'{path}/add_group_switch.json', 'w', Data)
+    utils.handle_json(f'{path}/add_group_switch.json', 'w', Data)
     await matcher.finish(MessageSegment.reply(event.message_id)+f"好~凌辉Bot已经{Temp}了本群的入群检测功能w~")
 
 handle_group = on_command("允许加群",aliases={"拒绝加群"},block=True)
